It's now possible to use a file backend for mail.
This allows much easier development and testing of that feature.
Could still use improvements though, currently what's being logged is
what would be sent to a function call serialized as json.

Other improvements were done such as avoid compiling the email template
because it is not a public page.

Also there was an issue that has been fixed: mustache as being used even
though we already switched to handlebars. A fix was done to use
handlebars instead, the template was renamed to reflect the change.

IMPORTANT: MUST VALIDATE WITH SMTP BACKEND. I have not tested for
regressions with previous commit, although there are no appearent issues
one never knows.