Changing internal structure which fixes compatibility with old versions.
Basically what changed is how the provider returns a message, it now
returns an Object which is muuuuuch easier to handle than a raw
formatted data. The responsability of correctely parsing and formatting
a message is now encapsulated on `parser.js` but should shortly be
refactored to use a better project structure.